Fixed table getting cut off

+++ Which Episode/ShowListTableViewController.m
@@ -198,12 +198,12 @@ const int Sort_ZtoA = 3;
     return 1;
 }
 
--(float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+-(CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     if([[UIDevice currentDevice] userInterfaceIdiom] == UIUserInterfaceIdiomPhone)
-        return 44.0f;
+        return 44.0;
     else
-        return 96.0f;
+        return 96.0 ;
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
